--- Braze Demo/View/ExpandableViewController/ExpandableViewController.xib
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="17506" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17505"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ExpandableViewController" customModule="Braze_Demo" customModuleProvider="target">
-            <connections>
-                <outlet property="imageHeightConstraint" destination="TxQ-zs-IhC" id="Nmt-Uf-MWp"/>
-                <outlet property="imageView" destination="PjC-6L-I4J" id="eKz-iM-x0c"/>
-                <outlet property="inAppMessageMessageLabel" destination="qBA-mV-LSS" id="cW8-RR-qvk"/>
-                <outlet property="view" destination="mzv-Pq-JQw" id="LfK-AB-DAi"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="mzv-Pq-JQw" customClass="ABKInAppMessageView">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="500"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <subviews>
-                <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="PjC-6L-I4J">
-                    <rect key="frame" x="25" y="485" width="364" height="0.0"/>
-                    <constraints>
-                        <constraint firstAttribute="height" id="TxQ-zs-IhC"/>
-                    </constraints>
-                </imageView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Lorem Ipsum" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="qBA-mV-LSS">
-                    <rect key="frame" x="157.5" y="69" width="99" height="20.5"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="8mF-Cc-fhJ">
-                    <rect key="frame" x="0.0" y="44" width="414" height="456"/>
-                    <connections>
-                        <action selector="sheetPressed:" destination="-1" eventType="touchUpInside" id="x23-xm-8O2"/>
-                    </connections>
-                </button>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="8yX-xp-bb7"/>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <gestureRecognizers/>
-            <constraints>
-                <constraint firstItem="qBA-mV-LSS" firstAttribute="top" secondItem="8yX-xp-bb7" secondAttribute="top" constant="25" id="Acs-rL-nFz"/>
-                <constraint firstItem="8yX-xp-bb7" firstAttribute="bottom" secondItem="8mF-Cc-fhJ" secondAttribute="bottom" id="Kfz-9E-blw"/>
-                <constraint firstItem="8mF-Cc-fhJ" firstAttribute="leading" secondItem="8yX-xp-bb7" secondAttribute="leading" id="Ws4-Jd-6k0"/>
-                <constraint firstItem="8mF-Cc-fhJ" firstAttribute="top" secondItem="8yX-xp-bb7" secondAttribute="top" id="e5p-SO-dRO"/>
-                <constraint firstItem="qBA-mV-LSS" firstAttribute="centerX" secondItem="8yX-xp-bb7" secondAttribute="centerX" id="eP4-Qa-5zv"/>
-                <constraint firstItem="PjC-6L-I4J" firstAttribute="centerX" secondItem="8yX-xp-bb7" secondAttribute="centerX" id="gbd-vE-8AP"/>
-                <constraint firstItem="PjC-6L-I4J" firstAttribute="top" relation="greaterThanOrEqual" secondItem="qBA-mV-LSS" secondAttribute="bottom" constant="10" id="joj-iw-Xav"/>
-                <constraint firstItem="8yX-xp-bb7" firstAttribute="trailing" secondItem="PjC-6L-I4J" secondAttribute="trailing" constant="25" id="mSV-c1-CsW"/>
-                <constraint firstItem="PjC-6L-I4J" firstAttribute="leading" secondItem="8yX-xp-bb7" secondAttribute="leading" constant="25" id="r2Q-Sh-u2f"/>
-                <constraint firstItem="8yX-xp-bb7" firstAttribute="trailing" secondItem="8mF-Cc-fhJ" secondAttribute="trailing" id="rzu-SQ-fea"/>
-                <constraint firstItem="8yX-xp-bb7" firstAttribute="bottom" secondItem="PjC-6L-I4J" secondAttribute="bottom" constant="15" id="xUE-PJ-gWg"/>
-            </constraints>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="166" y="100"/>
-        </view>
-    </objects>
-</document>
